% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{setCosts}
\alias{setCosts}
\alias{getCostOptions}
\title{Manipulate costs information}
\usage{
setCosts(
  composites,
  landclasses,
  outDir,
  outName,
  costOptions,
  costScns = NULL,
  extraProtectionValue = c(),
  exemptionClasses = c("urban", "cropland", "cutblock")
)

getCostOptions(costFile)
}
\arguments{
\item{composites}{TODO}

\item{landclasses}{TODO}

\item{outDir}{TODO}

\item{outName}{TODO}

\item{costOptions}{TODO}

\item{costScns}{TODO}

\item{extraProtectionValue}{TODO}

\item{exemptionClasses}{TODO}

\item{costFile}{TODO}
}
\value{
TODO
}
\description{
Two functions to manipulate the information related to costs
}
