% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels-utils.R
\name{getDistKernelFromMax}
\alias{getDistKernelFromMax}
\title{Get kernel distance from maximum}
\usage{
getDistKernelFromMax(kdim)
}
\arguments{
\item{kdim}{kernel dimensions.}
}
\value{
A matrix.
}
\description{
Calculates distance d from kernel dimensions.
}
