% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{uniformKernel}
\alias{uniformKernel}
\alias{exponentialKernel}
\title{Convolution kernels}
\usage{
uniformKernel(r = NULL, mu = NULL, cellDim = 1)

exponentialKernel(mu, cellDim = 1, negligible = 10^-10)
}
\arguments{
\item{cellDim}{TODO}

\item{negligible}{TODO}

\item{dmax}{TODO}

\item{useAveDist}{TODO}

\item{dbar}{}

\item{returnScale}{TODO}
}
\value{
A kernel (\code{matrix}).
}
\description{
Creates a convolution kernel
}
