% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastCostPathDistances.R
\name{leastCostPathDistances}
\alias{leastCostPathDistances}
\alias{leastCostPathDistances,RasterLayer-method}
\title{Least cost path distance matrix from rasters.}
\usage{
leastCostPathDistances(
  patches,
  cost,
  maxDist,
  bufferedPatches,
  neighbourhood = "octagon"
)

\S4method{leastCostPathDistances}{RasterLayer}(
  patches,
  cost,
  maxDist,
  bufferedPatches,
  neighbourhood = "octagon"
)
}
\arguments{
\item{patches}{RasterLayer. Binary patch map.}

\item{cost}{RasterLayer. Cost. Must be >0. Set cost=1 for euclidean distance.}

\item{maxDist}{Numeric. Maximum distance between i and j in map units.}

\item{bufferedPatches}{RasterLayer. Optional. If provided maxDist will be 
ignored. bufferedPatches defines the calculation region - use to restrict 
calculations to a portion of the map area.}

\item{neighbourhood}{Character. 'rook','queen', or 'octagon'. 'octagon' 
option is a modified version of the queen's 8 cell neighbourhood in which 
diagonals weights are 2^0.5x higher than horizontal/vertical weights.}
}
\value{
List of effective distance d_ij matrices for each subgraph.
}
\description{
Least cost path distance matrix from rasters.
}
\details{
TO DO: add description
}
\examples{
# TO DO: examples

}
