% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsBuffer.R
\docType{methods}
\name{lsBuffer}
\alias{lsBuffer}
\alias{lsBuffer,RasterLayer-method}
\title{Apply buffer.}
\usage{
lsBuffer(x, width, subset = NULL)

\S4method{lsBuffer}{RasterLayer}(x, width, subset = NULL)
}
\arguments{
\item{x}{RasterLayer.}

\item{width}{Numeric. Width of buffer to apply, in the units of x.}

\item{subset}{Logical expression. Optional. Passed as an argument to the subset function in order to select a subset of raster values or polygons. Default for rasters is expression(x!=0) - 0s and NAs are ignored. Default for polygons is NULL (all polygons selected).}
}
\value{
A binary RasterLayer object showing buffered area.
}
\description{
Apply buffer.
}
\details{
For a Raster* object use velox package.
}
\examples{
# TODO: examples
}
