% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTDConnect-package.R
\docType{package}
\name{LSTDConnect-package}
\alias{LSTDConnect}
\alias{LSTDConnect-package}
\title{LSTDConnect: ECCC Landscape Science & Technology Division Connectivity Analysis Tools}
\description{
Connectivity and distance weighted sum methods. This package is in development, and may or may not function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LandSciTech/LSTDConnect/}
  \item \url{https://landscitech.github.io/LSTDConnect/}
}

}
\author{
\strong{Maintainer}: Josie Hughes \email{josie.hughes@canada.ca}

Authors:
\itemize{
  \item Griffin Barrett \email{griffin-barrett@outlook.com}
  \item Valentin Lucet \email{valentin.lucet@gmail.com}
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Environment and Climate Change [copyright holder]
}

}
\keyword{internal}
