% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcConnectedness.R
\docType{methods}
\name{parcConnectedness}
\alias{parcConnectedness}
\alias{parcConnectedness,ParcConnectedness-method}
\alias{parcConnectedness,RasterStack-method}
\title{Calculate PARC-Connectedness}
\usage{
parcConnectedness(x, habitatValue = NULL,
  distanceMethod = "leastCostPaths", maxDist = 500, alpha = c(2, 25,
  100), metric = "connectivity", region = NULL, memoryLimit = 1,
  stopOnMemoryLimit = T)

\S4method{parcConnectedness}{ParcConnectedness}(x, habitatValue = NULL,
  distanceMethod = "leastCostPaths", maxDist = 500, alpha = c(2, 25,
  100), metric = "connectivity", region = NULL, memoryLimit = 1,
  stopOnMemoryLimit = T)

\S4method{parcConnectedness}{RasterStack}(x, habitatValue = NULL,
  distanceMethod = "leastCostPaths", maxDist = 500, alpha = c(2, 25,
  100), metric = "connectivity", region = NULL, memoryLimit = 1,
  stopOnMemoryLimit = T)
}
\arguments{
\item{x}{RasterStack, matrix or ParcConnectedness object. RasterStack of patches, cost, and habitatValue will be passed to distanceMethod for calculation of d_ij effective distance matrix. Or pass in precalculated effective distance matrix d_ij or ParcConnectedness object for increased computational efficiency.}

\item{habitatValue}{vector. Vector of H_j values. Indices must correspond to rows(?) of d_ij matrix x. Ignored if x is a RasterStack or ParcConnectedness object.}

\item{distanceMethod}{Character string. Options are leastCostPaths or ... Argument is ignored if x is a matrix or ParcConnectedness object.}

\item{maxDist}{Number. Maximum distance between i and j in units of RasterStack x. Ignored if x is a matrix or ParcConnectedness object.}

\item{alpha}{Number or vector of these. 1/alpha is in units of effective distance. See Drielsma et al 2007 for details.}

\item{metric}{Character string. Options are "connectivity" or "colonization potential" (equations 3 or 4 of Drielsma)}

\item{region}{RasterLayer or character string. Optional. Areas to treat as distinct in calculations. If x is a ParcConnectedness object region should be one or more names of the d_ij list.}

\item{memoryLimit}{Number. Total number of GB to allocate to d_ij list. When size of d_ij list exceeds this limit calculations will be done, but matrices will not be retained and returned.}

\item{stopOnMemoryLimit}{Boolean. If true algorithm will return error if expected memory requirement is greater than available memory.}
}
\value{
A ParcConnectedness object including patches, effective distance d_ij matrix, habitat value H_j vector, M_i maps, average metric value for each patch Mbar_p, data frame average metric of all patches Mbar.
}
\description{
Calculate PARC-Connectedness
}
\details{
TO DO: add description
Average colonization potential will be calculated for each stratum (or patch id) in patches layer.

cost must be >0. Set cost = 1 for euclidean distance.
maxDist is in map units (metres etc). See leastCostPathDistances() for details.

Algorithm notes:
  The limiting step is calculation and storage of d_ij matrix the set of pixels within a buffered clump.
  Given N protected pixels in a clump, and B pixels within maxDist of the protected pixels, the size of d_ij will be N*B.
  For example, d_ij for a circular park with 10km diameter, maxDist=500km, 1km resolution, cost>=1: B~800000, N~80, N*B~63 million.
  Obviously we will need to do something about this for maps with big park clusters - it works fine for small examples.
}
\examples{
# TO DO: examples
#TO DO: define parcConnectedness for other input signatures.
}
