% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAHelperFunctions.R
\name{roadCLUS.getGraph}
\alias{roadCLUS.getGraph}
\title{TODO}
\usage{
roadCLUS.getGraph(sim, neighbourhood)
}
\arguments{
\item{sim}{TODO}

\item{neighbourhood}{TODO}
}
\value{
TODO
}
\description{
TODO
}
