% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{samc}
\alias{samc}
\alias{distribution}
\title{Spatial Absorbing Markov Chain (SAMC) in parallel}
\usage{
samc(
  resistance,
  absorption = NULL,
  fidelity = NULL,
  directions = 8,
  kernel = NULL,
  resistance_na_mask = 0,
  absorption_na_mask = 0,
  fidelity_na_mask = 0,
  symmetric = TRUE
)

distribution(samc, occ, time = 1, dead = NULL)
}
\arguments{
\item{resistance}{Deprecated. Use the \code{data} parameter.}

\item{absorption}{A \code{\link[raster]{RasterLayer-class}} or \code{\link[base]{matrix}}}

\item{fidelity}{A \code{\link[raster]{RasterLayer-class}} or \code{\link[base]{matrix}}}

\item{directions}{Either 4 (rook) or 8 (queen), the directions of movement.}

\item{kernel}{Alternative to "directions", a square kernel.}

\item{resistance_na_mask}{Default to 0, the value to give to NAs and NaNs values
in resistance.}

\item{absorption_na_mask}{Default to 0, the value to give to NAs and NaNs
values in absorption.}

\item{fidelity_na_mask}{Default to 0, the value to give to NAs and NaNs
values in fidelity.}

\item{symmetric}{Default to true, whether movement is symmetrical
between cells.}

\item{samc}{A list resulting from calling samc.}

\item{occ}{The initial state of the model (abundance matrix).}

\item{time}{A positive integer or a vector of positive integers representing
time steps.}

\item{dead}{Optionnal matrix containing information on deaths.}
}
\value{
\code{samc} will return a list.
\code{distribution} will also return a list.
}
\description{
Implements the SAMC algorithm from \link[samc:samc]{samc}.
}
