% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{uniformKernel}
\alias{uniformKernel}
\alias{exponentialKernel}
\title{Convolution kernels}
\usage{
uniformKernel(dmax, cellDim = 1, useAveDist = F)

exponentialKernel(
  dbar,
  cellDim = 1,
  negligible = 10^-10,
  returnScale = F,
  dmax = NULL
)
}
\arguments{
\item{dmax}{TODO}

\item{cellDim}{TODO}

\item{useAveDist}{TODO}

\item{dbar}{TODO}

\item{negligible}{TODO}

\item{returnScale}{TODO}
}
\value{
A kernel (`matrix`).
}
\description{
Creates a convolution kernel
}
