% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{uniformKernel}
\alias{uniformKernel}
\alias{exponentialKernel}
\title{Convolution kernels}
\usage{
uniformKernel(r = NULL, mu = NULL, cellDim = 1)

exponentialKernel(mu, cellDim = 1, negligible = 10^-10)
}
\arguments{
\item{r}{The radius of the kernel, in physical units.}

\item{mu}{Alternative to r, so that r = 3 * mu / 2 (equation for the
average distance from the center of a circle).}

\item{cellDim}{Default to 1, cell dimensions used as a scaling factor
for the kernel size}

\item{negligible}{Negligible value when rounding edges of exponential kernels.}
}
\value{
A kernel (\code{matrix}).
}
\description{
Creates a convolution kernel
}
