% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{samc_cache}
\alias{samc_cache}
\alias{samc_step}
\title{SAMC Implementation}
\usage{
samc_cache(
  resistance,
  fidelity = NULL,
  absorbtion = NULL,
  kernel = 8,
  resistance_na_mask = 0,
  absorbtion_na_mask = 0,
  fidelity_na_mask = 0,
  symmetric = TRUE
)

samc_step(steps = 1, cache, population, dead = NULL)
}
\arguments{
\item{resistance}{TODO}

\item{fidelity}{TODO}

\item{absorbtion}{TODO}

\item{kernel}{TODO}

\item{resistance_na_mask}{TODO}

\item{absorbtion_na_mask}{TODO}

\item{fidelity_na_mask}{TODO}

\item{symmetric}{TODO}

\item{steps}{TODO}

\item{cache}{TODO}

\item{population}{TODO}

\item{dead}{TODO}
}
\value{
TODO
}
\description{
TODO
}
